Refactor out Header component and unify styled-components and css usage
