+++ src/App.scss
@@ -7,18 +7,6 @@ $navButtonHeight: 72px;
   flex-flow: column nowrap;
   flex: 1;
 }
-.nav {
-  display: flex;
-  justify-content: space-between;
-  text-align: center;
-  width: 100%;
-  > div {
-    width: 30vw;
-    height: $navButtonHeight;
-    // COLORS.backgroundPrimary
-    background-color: #626262;
-  }
-}
 
 .statsSection {
   display: flex;
@@ -34,7 +22,7 @@ $navButtonHeight: 72px;
   width: 100%;
   display: flex;
   flex-flow: row wrap;
-  flex: 1 0 auto;
+  flex: 1 1 auto;
   align-content: stretch;
 
   &.powers {
