+++ src/util/layout.ts
@@ -1,0 +1,10 @@
+const LAYOUT = {
+  // Height of header in px
+  navHeight: 72,
+  // Height of talents section in px
+  talentsHeight: 80,
+  // Height of footer in px
+  footerHeight: 94,
+};
+
+export default LAYOUT;
